% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{add_linear_regression_formula}
\alias{add_linear_regression_formula}
\title{Add Linear Regression Formula (not currently used)}
\usage{
add_linear_regression_formula(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
A ggplot object with the formula placed on the top
}
\description{
A function that adds linear regression formula and place it as a text annotation
on the top of a plot. Requires ggplot object as input.
}
