% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binary_cat_dist}
\alias{binary_cat_dist}
\title{Function to generate a binary categorical covariate}
\usage{
binary_cat_dist(n = 20, percent = 50, catvalue1 = 1, catvalue2 = 0)
}
\arguments{
\item{n}{number of subjects}

\item{percent}{percent of subjects (approximate) in first category}

\item{catvalue1}{numeric value for first category}

\item{catvalue2}{numeric value for second category}
}
\value{
a numeric vector
}
\description{
Function to generate a binary categorical covariate
}
