% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{calc_summary_stats}
\alias{calc_summary_stats}
\title{Calculate summary stats with optional rounding}
\usage{
calc_summary_stats(
  orig_data,
  dp = 1,
  sigdig = FALSE,
  convert_to_numeric = TRUE,
  transpose = FALSE,
  check_empty_rows = TRUE,
  id_colname = "ID",
  comma_format = TRUE
)
}
\arguments{
\item{orig_data}{input dataframe}

\item{dp}{round to this many decimal places}

\item{sigdig}{Set to TRUE to round using significant digits instead}

\item{convert_to_numeric}{Default TRUE, convert df to numeric and turns characters to NA's}

\item{transpose}{Set to TRUE to transpose the table such that each variable
(column) becomes a row}

\item{check_empty_rows}{Default TRUE, will not perform summary stats if there are no rows in df}

\item{id_colname}{The ID column to distinct by and then removed before summary calcs are done}

\item{comma_format}{Set to TRUE to use big mark formatting}
}
\value{
a dataframe with summary stats
}
\description{
Calculate summary stats with optional rounding
}
