% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_10.R
\name{calc_summary_stats_as_list}
\alias{calc_summary_stats_as_list}
\title{Return summary stats as a list}
\usage{
calc_summary_stats_as_list(
  nca_df,
  group_by_name,
  list_of_nca_metrics = c("AUCLST", "AUCIFO", "CMAX", "CMAXD", "LAMZHL", "TMAX",
    "MRTIVLST", "MRTIVIFO", "MRTIVIFP", "MRTEVLST", "MRTEVIFO", "MRTEVIFP", "VZO", "VZP",
    "VZFO", "VZFP", "CLO", "CLP", "CLFO", "CLFP"),
  dp = 3,
  sigdig = TRUE,
  convert_to_numeric = FALSE,
  transpose = TRUE,
  id_colname = "ID"
)
}
\arguments{
\item{nca_df}{input dataframe (after using NonCompart::tblNCA)}

\item{group_by_name}{the singular "key" argument for NonCompart::tblNCA \emph{minus} the ID column}

\item{list_of_nca_metrics}{columns to retain in the summary table}

\item{dp}{round to this many decimal places}

\item{sigdig}{Set to TRUE to round using significant digits instead}

\item{convert_to_numeric}{Default FALSE, convert df to numeric and turns characters to NA's}

\item{transpose}{Set to TRUE to transpose the table such that each variable
(column) becomes a row}

\item{id_colname}{The ID column to distinct by and then removed before summary calcs are done}
}
\value{
a list containing X number of dataframes for each unique value of group_by_name
}
\description{
Return summary stats as a list
}
