% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Function to split a continuous X-variable for a number of quantiles}
\usage{
calculate_quantiles(df, xvar, num_quantiles)
}
\arguments{
\item{df}{Name of dataframe}

\item{xvar}{Name of x-axis variable to split by}

\item{num_quantiles}{Number of discrete quantiles}
}
\value{
a dataframe containing the quantile limits
}
\description{
Function to split a continuous X-variable for a number of quantiles
}
