% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{categorize_outliers}
\alias{categorize_outliers}
\title{Function that categorizes outliers relative to a supplied stratification
group}
\usage{
categorize_outliers(df, highlight_range, y_axis, strat_by, debug = FALSE)
}
\arguments{
\item{df}{Name of input dataframe}

\item{highlight_range}{A character string in percentage e.g. "80\%" to define outlier threshold}

\item{y_axis}{Y-axis to derive the arithmetic mean for ID and group}

\item{strat_by}{Column name to be used as a stratification variable for group}

\item{debug}{Show debugging messages}
}
\value{
a dataframe with the new column called "outlier_status", with three categories: "Above", "Below", "Within"
}
\description{
Function that categorizes outliers relative to a supplied stratification
group
}
