% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_12.R
\name{check_and_combine_df}
\alias{check_and_combine_df}
\title{combined 2 datasets for download}
\usage{
check_and_combine_df(
  model_1_is_valid,
  model_2_is_valid,
  input_df_1 = NULL,
  input_df_2 = NULL
)
}
\arguments{
\item{model_1_is_valid}{logical TRUE/FALSE}

\item{model_2_is_valid}{logical TRUE/FALSE}

\item{input_df_1}{Model 1 simulated data}

\item{input_df_2}{Model 2 simulated data}
}
\value{
a df
}
\description{
combined 2 datasets for download
}
