% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{check_cov_name_duplicate}
\alias{check_cov_name_duplicate}
\title{Function to check covariate names do not duplicate}
\usage{
check_cov_name_duplicate(current_id, all_ids)
}
\arguments{
\item{current_id}{: id of current textInput}

\item{all_ids}{: list of ids for textInputs to check}
}
\description{
Function to check covariate names do not duplicate
}
