% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{check_matrix}
\alias{check_matrix}
\title{Checks for validity of matrix}
\usage{
check_matrix(
  input_model_object,
  input_matrix,
  check_diagonal = TRUE,
  check_symmetry = FALSE,
  debug = FALSE,
  display_error = TRUE
)
}
\arguments{
\item{input_model_object}{mrgmod object}

\item{input_matrix}{a list where each element is a matrix}

\item{check_diagonal}{Default TRUE, check that each diagonal must be >= 0}

\item{check_symmetry}{Default FALSE, check for symmetry}

\item{debug}{Show debugging messages}

\item{display_error}{Displays a showNotification message pop-up}
}
\value{
logical (TRUE/FALSE), where TRUE means pass (valid matrix)
}
\description{
Checks for validity of matrix
}
