% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_8.R
\name{convert_to_plotly_watermark}
\alias{convert_to_plotly_watermark}
\title{Converts a ggplot object to plotly object and then apply watermark, and other options}
\usage{
convert_to_plotly_watermark(
  ggplot_object,
  opacity = 0.1,
  font_name = "Arial",
  format = "png",
  filename = "newplot",
  width = NULL,
  height = NULL,
  debug = FALSE,
  plotly_watermark = TRUE
)
}
\arguments{
\item{ggplot_object}{a ggplot object}

\item{opacity}{the alpha or transparency, goes from 0 to 1}

\item{font_name}{font name as a string}

\item{format}{one of "png", "svg", "jpeg", "webp"}

\item{filename}{name of output file when saved}

\item{width}{width in pixels. Default NULL - uses current resolution}

\item{height}{height in pixels. Default NULL - uses current resolution}

\item{debug}{Set to TRUE to show debug messages}

\item{plotly_watermark}{Set to TRUE to insert "For Internal Use Only"}
}
\value{
a plotly object
}
\description{
Converts a ggplot object to plotly object and then apply watermark, and other options
}
