% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{create_alert}
\alias{create_alert}
\title{Function that generates a shinyalert}
\usage{
create_alert(
  ppm_name = "Firstname Lastname",
  ppm_email = "dummy.email@company.com"
)
}
\arguments{
\item{ppm_name}{Name of Project Pharmacometrician}

\item{ppm_email}{Email of Project Pharmacometrician}
}
\value{
a shinyalert() popup
}
\description{
Function that generates a shinyalert
}
