% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_facet_label}
\alias{create_facet_label}
\title{Function that combines multiple columns to create a new one that contains
the facet labels}
\usage{
create_facet_label(df, sort_by)
}
\arguments{
\item{df}{Name of input dataframe}

\item{sort_by}{A list containing columns to create label by. Uses first row only}
}
\value{
a dataframe with the new column called "facet_label"
}
\description{
Function that combines multiple columns to create a new one that contains
the facet labels
}
