% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_8.R
\name{create_value_box}
\alias{create_value_box}
\title{Short-hand for creating NCA metrics valueBoxes}
\usage{
create_value_box(
  input_dataset,
  name_ends_with,
  value_box_subtitle,
  width = infoBox_width,
  color,
  sigdig = 4,
  dp = FALSE
)
}
\arguments{
\item{input_dataset}{dataset to extract metrics from}

\item{name_ends_with}{name of metrics column that ends with x (string)}

\item{value_box_subtitle}{name of subtitle in value box}

\item{width}{width of valueBox}

\item{color}{color of valueBox}

\item{sigdig}{significant digits to output}

\item{dp}{if TRUE, uses decimal rounding instead}
}
\value{
a shinydashboard valueBox
}
\description{
Short-hand for creating NCA metrics valueBoxes
}
