% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_data_page_ind_plot}
\alias{do_data_page_ind_plot}
\title{Individual Plot for Data Exploration}
\usage{
do_data_page_ind_plot(
  nmd,
  rownums,
  colnums,
  pagenum = 1,
  filter_id,
  filter_cmt,
  sort_by,
  strat_by,
  highlight_range,
  x_axis,
  y_axis,
  color_by,
  med_line,
  med_line_by,
  boxplot,
  dolm,
  smoother,
  facet_name = "ID",
  logy,
  lby = logbreaks_y,
  lbx = logbreaks_x,
  logx,
  plot_title,
  label_size = 3,
  discrete_threshold = 7,
  boxplot_x_threshold = 20,
  error_text_color = "#F8766D",
  highlight_var,
  highlight_var_values,
  plot_dosing = TRUE,
  same_scale = FALSE,
  dose_col = "DOSE",
  dose_units,
  lloq_name = "",
  debug = FALSE
)
}
\arguments{
\item{nmd}{The NONMEM dataset for plotting (requires ID, TIME, DV at minimum)}

\item{rownums}{How many rows per page}

\item{colnums}{How many cols per page}

\item{filter_id}{Filter by these IDs}

\item{filter_cmt}{Filter by this CMT}

\item{sort_by}{Sort by one or more variables (a list)}

\item{strat_by}{Stratify by 1 variable and highlight outliers}

\item{highlight_range}{Flag potential outliers when they are higher or below this mean value of the group, a string}

\item{x_axis}{X-axis for plot}

\item{y_axis}{Y-axis for plot}

\item{color_by}{Color by this column}

\item{med_line}{When TRUE, will draw median line by equidistant X-axis bins}

\item{med_line_by}{Column name to draw median line by}

\item{boxplot}{Draws a geom_boxplot instead of geom_point and geom_line}

\item{dolm}{Insert linear regression with formula on top of plot}

\item{smoother}{Insert smoother}

\item{facet_name}{variable to facet by}

\item{logy}{Log Y-axis}

\item{lby}{Log breaks for Y-axis}

\item{lbx}{Log breaks for X-axis}

\item{logx}{Log X-axis}

\item{plot_title}{Optional plot title}

\item{label_size}{font size for geom_text labels (N=x for boxplots or linear regressions)}

\item{discrete_threshold}{Draws geom_count if there are <= this number of unique Y-values}

\item{boxplot_x_threshold}{Throws an error when the unique values of x-axis exceeds this number}

\item{error_text_color}{error text color for element text}

\item{highlight_var}{variable name to be highlighted by a different shape}

\item{highlight_var_values}{variable values associated with highlight_var to be highlighted}

\item{plot_dosing}{Plots dosing line and dose text when TRUE}

\item{same_scale}{Uses fixed scales based on entire dataset when TRUE}

\item{dose_col}{name of dose column, usually AMT or DOSE}

\item{dose_units}{(Optional) name of dose units, usually mg or nmol}

\item{lloq_name}{(Optional) Supply name of lloq column to be plotted as hline}

\item{debug}{show debugging messages}
}
\value{
a ggplot object
}
\description{
This is the main function for plotting individual plots from uploaded datasets.
It is designed to be flexible enough to handle continuous/continuous, discrete/continuous, and
discrete/discrete type of data to cover most use cases.
}
