% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_9.R
\name{do_data_page_plot}
\alias{do_data_page_plot}
\title{Quick Plot for Data Exploration}
\usage{
do_data_page_plot(
  nmd,
  filter_cmt,
  x_axis,
  y_axis,
  color_by,
  med_line,
  med_line_by,
  dolm,
  smoother,
  facet_name,
  logy,
  lby = logbreaks_y,
  lbx = logbreaks_x,
  logx,
  plot_title,
  debug = FALSE
)
}
\arguments{
\item{nmd}{The NONMEM dataset for plotting (requires ID, TIME, DV at minimum)}

\item{filter_cmt}{Filter by this CMT}

\item{x_axis}{X-axis for plot}

\item{y_axis}{Y-axis for plot}

\item{color_by}{Color by this column}

\item{med_line}{When TRUE, will draw median line by equidistant X-axis bins}

\item{med_line_by}{Column name to draw median line by}

\item{dolm}{Insert linear regression with formula on top of plot}

\item{smoother}{Insert smoother}

\item{facet_name}{variable to facet by}

\item{logy}{Log Y-axis}

\item{lby}{Log breaks for Y-axis}

\item{lbx}{Log breaks for X-axis}

\item{logx}{Log X-axis}

\item{plot_title}{Optional plot title}

\item{debug}{show debugging messages}
}
\value{
a ggplot object
}
\description{
Quick Plot for Data Exploration
}
