% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_12.R
\name{draw_correlation_plot}
\alias{draw_correlation_plot}
\title{Draw a correlation plot from a NONMEM-formatted dataset}
\usage{
draw_correlation_plot(
  input_df,
  corr_variables,
  color_sep = "",
  catcov_threshold = 10,
  debug = FALSE
)
}
\arguments{
\item{input_df}{Input dataframe}

\item{corr_variables}{Vector (string) of names to be used in the plot}

\item{color_sep}{Variable (string) name used for colour separator}

\item{catcov_threshold}{assumes a covariate is categorical if the unique values
in the covariate are less than this threshold (default 10),
should be less than nsubj that is available from the dataset}

\item{debug}{show debugging messages}
}
\value{
a ggplot object
}
\description{
Draw a correlation plot from a NONMEM-formatted dataset
}
