% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_histogram_plot}
\alias{draw_histogram_plot}
\title{Draw a histogram plot from a NONMEM-formatted dataset for covariates}
\usage{
draw_histogram_plot(input_df, hist_variables, bin_size = 5, debug = FALSE)
}
\arguments{
\item{input_df}{Input dataframe}

\item{hist_variables}{Vector (string) of names to be used in the plot}

\item{bin_size}{Size of bin_width used for histogram}

\item{debug}{show debugging messages}
}
\value{
a ggplot object
}
\description{
Draw a histogram plot from a NONMEM-formatted dataset for covariates
}
