% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{eval_parse}
\alias{eval_parse}
\title{Short-hand function to eval, and then parse a string}
\usage{
eval_parse(text)
}
\arguments{
\item{text}{Some text}
}
\value{
A list with two elements:
\itemize{
\item \code{result}: The result of \code{eval(parse(text))}, or \code{NULL} if an error occurred.
\item \code{error}: The error that occurred, or \code{NULL} if no error occurred.
}
}
\description{
Short-hand function to eval, and then parse a string
}
