% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{exposures_table}
\alias{exposures_table}
\title{Function to calculate exposures from IIV output}
\usage{
exposures_table(
  input_simulated_table,
  output_conc,
  start_time = NULL,
  end_time = NULL,
  debug = FALSE
)
}
\arguments{
\item{input_simulated_table}{a dataframe (usually created from mrgsim)}

\item{output_conc}{y variable of interest to perform summary stats calc}

\item{start_time}{start time interval for metrics}

\item{end_time}{end time interval for metrics}

\item{debug}{show debugging messages}
}
\value{
a dataframe with additional columns of summary stats
}
\description{
Function to calculate exposures from IIV output
}
