% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{extract_matrix}
\alias{extract_matrix}
\title{Function to output a matrix from mrgsolve model object}
\usage{
extract_matrix(
  input_model_object,
  name_of_matrix = "omega",
  remove_upper_tri = TRUE,
  remove_zero = TRUE,
  coerce_to_character = TRUE,
  debug = show_debugging_msg
)
}
\arguments{
\item{input_model_object}{mrgsolve model object to extract from}

\item{name_of_matrix}{either "omega" or "sigma", will use "omat" or "smat" accordingly}

\item{remove_upper_tri}{When TRUE, turns upper triangle of matrix to NA_real_}

\item{remove_zero}{In case any of them are 0, don't replace the diagonals}

\item{coerce_to_character}{Default TRUE, turns all output into a character as a workaround to prevent rounding}

\item{debug}{When TRUE, show debugging messages}
}
\value{
a matrix
}
\description{
Function to output a matrix from mrgsolve model object
}
