% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{extract_model_params}
\alias{extract_model_params}
\title{Function to extract params from model object}
\usage{
extract_model_params(input_model_object)
}
\arguments{
\item{input_model_object}{Expects a mrgsolve model object}
}
\value{
a dataframe of all names inside $PARAM where each row contains
a parameter name in column 1, and parameter value in column 2
}
\description{
Function to extract params from model object
}
\note{
Should be a tibble of 2 columns and X rows where X = number of params
}
