% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{general_warning_modal}
\alias{general_warning_modal}
\title{Creates a general warning modal}
\usage{
general_warning_modal(title = "Warning", text_description = "Test")
}
\arguments{
\item{title}{Title for the modal box}

\item{text_description}{UI elements}
}
\value{
a modal dialog
}
\description{
Creates a general warning modal
}
