% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{generate_dosing_regimens}
\alias{generate_dosing_regimens}
\title{Generate Dosing Regimens}
\usage{
generate_dosing_regimens(
  amt1,
  delay_time1,
  cmt1,
  tinf1,
  total1,
  ii1,
  amt2,
  delay_time2,
  cmt2,
  tinf2,
  total2,
  ii2,
  amt3,
  delay_time3,
  cmt3,
  tinf3,
  total3,
  ii3,
  amt4,
  delay_time4,
  cmt4,
  tinf4,
  total4,
  ii4,
  amt5,
  delay_time5,
  cmt5,
  tinf5,
  total5,
  ii5,
  mw_conversion = FALSE,
  wt_multiplication_value = 1,
  create_dummy_ev = TRUE,
  debug = FALSE
)
}
\arguments{
\item{amt1}{Dose Amount Regimen 1}

\item{delay_time1}{Delay Time Regimen 1}

\item{cmt1}{Input CMT Regimen 1}

\item{tinf1}{Infusion Time Regimen 1}

\item{total1}{Total Doses Regimen 1}

\item{ii1}{Interdose Interval Regimen 1}

\item{amt2}{Dose Amount Regimen 2}

\item{delay_time2}{Delay Time Regimen 2}

\item{cmt2}{Input CMT Regimen 2}

\item{tinf2}{Infusion Time Regimen 2}

\item{total2}{Total Doses Regimen 2}

\item{ii2}{Interdose Interval Regimen 2}

\item{amt3}{Dose Amount Regimen 3}

\item{delay_time3}{Delay Time Regimen 3}

\item{cmt3}{Input CMT Regimen 3}

\item{tinf3}{Infusion Time Regimen 3}

\item{total3}{Total Doses Regimen 3}

\item{ii3}{Interdose Interval Regimen 3}

\item{amt4}{Dose Amount Regimen 4}

\item{delay_time4}{Delay Time Regimen 4}

\item{cmt4}{Input CMT Regimen 4}

\item{tinf4}{Infusion Time Regimen 4}

\item{total4}{Total Doses Regimen 4}

\item{ii4}{Interdose Interval Regimen 4}

\item{amt5}{Dose Amount Regimen 5}

\item{delay_time5}{Delay Time Regimen 5}

\item{cmt5}{Input CMT Regimen 5}

\item{tinf5}{Infusion Time Regimen 5}

\item{total5}{Total Doses Regimen 5}

\item{ii5}{Interdose Interval Regimen 5}

\item{mw_conversion}{MW conversion}

\item{wt_multiplication_value}{Weight multiplication value}

\item{create_dummy_ev}{Default TRUE, to create dummy ev if there are no valid
dose amounts, set to FALSE to not create one}

\item{debug}{set to TRUE to show debug messages}
}
\value{
A mrgsolve::ev event object
}
\description{
This is the main function to supply dosing regimens to simulations.
If there are no valid dose amounts from any dosing regimens, a dummy mrgsolve::ev()
object is created to supply a dose of 0 to be used for simulations.
}
