% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{generate_log_breaks}
\alias{generate_log_breaks}
\title{Generating log breaks and axis labels, mainly for plotly}
\usage{
generate_log_breaks(base_values, start, end)
}
\arguments{
\item{base_values}{the base value(s) to apply the exponent to}

\item{start}{the starting exponent}

\item{end}{the last exponent}
}
\value{
a numeric
}
\description{
Generating log breaks and axis labels, mainly for plotly
}
