% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_api_key_name}
\alias{get_api_key_name}
\title{Returns the designated API key name from a provider in .Renviron File}
\usage{
get_api_key_name(llm_service)
}
\arguments{
\item{llm_service}{Name of LLM provider}
}
\value{
A character string containing the API key name
}
\description{
Retrieves the name of the API key for a specified LLM service from environment variables.
Set environment variables with \code{usethis::edit_r_environ()}.
}
