% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric}

\item{na.rm}{Set to TRUE to remove NAs}
}
\value{
A numeric
}
\description{
Geometric mean
}
