% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gm_mean_cv}
\alias{gm_mean_cv}
\title{Geometric mean CV \%}
\usage{
gm_mean_cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric}

\item{na.rm}{Set to TRUE to remove NAs}
}
\value{
A numeric
}
\description{
Geometric mean CV \%
}
