% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{handle_blanks}
\alias{handle_blanks}
\title{Function that turns blank values into ".(blanks)"
group}
\usage{
handle_blanks(df, column_name)
}
\arguments{
\item{df}{Input dataframe}

\item{column_name}{Name of column (must be character type) to check for blanks}
}
\value{
a dataframe with the blank values converted into ".(blanks)
}
\description{
Function that turns blank values into ".(blanks)"
group
}
