% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iterate_batch_runs}
\alias{iterate_batch_runs}
\title{Perform multiple simulations from a dataframe containing parameters}
\usage{
iterate_batch_runs(
  batch_run_df,
  input_model_object,
  pred_model = FALSE,
  wt_based_dosing = FALSE,
  wt_name = "WT",
  ev_df,
  model_dur = FALSE,
  model_rate = FALSE,
  sampling_times,
  divide_by = 1,
  debug = FALSE,
  append_id_text = "m1-",
  show_progress = TRUE,
  gradient = FALSE,
  nsubj_offset = 0,
  parallel_sim = FALSE,
  parallel_n = 200
)
}
\arguments{
\item{batch_run_df}{Input batch run dataframe containing "Name", "Reference", "Lower", "Upper"}

\item{input_model_object}{mrgmod object}

\item{pred_model}{Default FALSE. set to TRUE to set ev_df CMT to 0}

\item{wt_based_dosing}{When TRUE, will try to use weight-based dosing}

\item{wt_name}{Name of weight parameter in model to multiply dose amt by}

\item{ev_df}{ev() dataframe containing dosing info}

\item{model_dur}{Default FALSE, set to TRUE to model duration inside the code}

\item{model_rate}{Default FALSE, set to TRUE to model rate inside the code}

\item{sampling_times}{A vector of sampling times (note: not a tgrid object)}

\item{divide_by}{Divide the TIME by this value, used for scaling x-axis}

\item{debug}{When TRUE, outputs debugging messages}

\item{append_id_text}{A string prefix to be inserted for each ID}

\item{show_progress}{When TRUE, shows shiny progress messages}

\item{gradient}{When TRUE, perform gradient runs in between lower/upper}

\item{parallel_sim}{Default FALSE, uses the future and mrgsim.parallel packages !Not implemented live!}

\item{parallel_n}{The number of subjects required before parallelization is used !Not implemented live!}
}
\value{
a df mrgsolve output totaling 2 * params + 1 (reference) runs, or 8 * params + 1 if gradient option is used
}
\description{
Perform multiple simulations from a dataframe containing parameters
}
