% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Linear regression text (for plotly)}
\usage{
lm_eqn(df, facet_name, x, y)
}
\arguments{
\item{df}{A dataframe}

\item{facet_name}{Optional column name (string) to facet by}

\item{x}{column name (string) of x-variable}

\item{y}{column name (string) of y-variable}
}
\value{
A dataframe containing the column "label" containing the string
formula of R2 coefficient for use in ggplot labels or title
}
\description{
Linear regression text (for plotly)
}
