% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lm_eqn_old}
\alias{lm_eqn_old}
\title{Linear regression text (not used)}
\usage{
lm_eqn_old(df, x, y)
}
\arguments{
\item{df}{A dataframe}

\item{x}{column name (string) of x-variable}

\item{y}{column name (string) of y-variable}
}
\value{
A character string containing the expression of the R2 coefficient
for use in ggplot labels or title
}
\description{
Linear regression text (not used)
}
