% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_10.R
\name{lowerFn}
\alias{lowerFn}
\title{Function to support draw_correlation_plot}
\usage{
lowerFn(data, mapping, method = "lm", ...)
}
\arguments{
\item{data}{input df for ggplot}

\item{mapping}{mapping for ggplot}

\item{method}{plotting method for ggplot (default "lm")}

\item{...}{other parameters to pass onto geoms}
}
\value{
a ggplot object
}
\description{
Function to support draw_correlation_plot
}
