% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\name{model_code_2cmt_MM_TMDD}
\alias{model_code_2cmt_MM_TMDD}
\title{MM approximation of TMDD, see Dua et al (2015)
https://doi.org/10.1002/psp4.41}
\usage{
model_code_2cmt_MM_TMDD
}
\description{
MM approximation of TMDD, see Dua et al (2015)
https://doi.org/10.1002/psp4.41
}
