% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\name{model_examples_list}
\alias{model_examples_list}
\title{List of template models available for all users}
\usage{
model_examples_list
}
\description{
List of template models available for all users
}
