% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MVPbuildTooltipOrPopoverOptionsList}
\alias{MVPbuildTooltipOrPopoverOptionsList}
\title{Exported from shinyBS:::buildTooltipOrPopoverOptionsList}
\usage{
MVPbuildTooltipOrPopoverOptionsList(
  title,
  placement,
  trigger,
  options,
  content
)
}
\description{
Exported from shinyBS:::buildTooltipOrPopoverOptionsList
}
\keyword{internal}
