% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{pct_above_y_at_x}
\alias{pct_above_y_at_x}
\title{Function that calculates \% of population > Y-value at
a given X-value}
\usage{
pct_above_y_at_x(
  model_is_valid = FALSE,
  input_df,
  y_name = "DV",
  y_value = NA,
  x_name = "TIME",
  x_value = NA,
  return_number_ids = FALSE
)
}
\arguments{
\item{model_is_valid}{checkpoint before rest of function proceeds, default FALSE}

\item{input_df}{Dataset containing ID column with more than 1 ID}

\item{y_name}{Name of column (string) for Y-value}

\item{y_value}{Y-value threshold (>), e.g. plasma concentration}

\item{x_name}{Name of column (string) for X-value}

\item{x_value}{X-value threshold (==), e.g. time}

\item{return_number_ids}{Return number of IDs that fits the criteria instead of \%
Default FALSE}
}
\value{
a numeric (either as a proportion (from 0 to 1, where 0 means 0\% and
1 means 100\%), or the actual number of IDs)
}
\description{
Function that calculates \% of population > Y-value at
a given X-value
}
