% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pdfNCA_wm}
\alias{pdfNCA_wm}
\title{Modified ncar::pdfNCA to support watermarks}
\usage{
pdfNCA_wm(
  fileName = "Temp-NCA.pdf",
  concData,
  key = "Subject",
  colTime = "Time",
  colConc = "conc",
  dose = 0,
  adm = "Extravascular",
  dur = 0,
  doseUnit = "mg",
  timeUnit = "h",
  concUnit = "ug/L",
  down = "Linear",
  R2ADJ = 0,
  MW = 0,
  SS = FALSE,
  iAUC = "",
  excludeDelta = 1,
  watermark = TRUE,
  internal_version = TRUE,
  debug_msg = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{fileName}{file name to save}

\item{concData}{concentration data table}

\item{key}{column names of concData to be shown in the output table}

\item{colTime}{column name for time}

\item{colConc}{column name for concentration}

\item{dose}{administered dose, a scalar or a vector}

\item{adm}{one of \code{"Bolus"} or \code{"Infusion"} or \code{"Extravascular"} to indicate drug administration mode}

\item{dur}{duration of infusion, a scalar or a vector}

\item{doseUnit}{unit of dose}

\item{timeUnit}{unit of time}

\item{concUnit}{unit of concentration}

\item{down}{either of \code{"Linear"} or \code{"Log"} to indicate the way to calculate AUC and AUMC}

\item{R2ADJ}{Minimum adjusted R-square value to determine terminal slope automatically}

\item{MW}{molecular weight of drug}

\item{SS}{if steady-state, this should be TRUE. AUCLST (AUClast) is used instead of AUCIFO (AUCinf) for the calculation of Vz (VZFO, VZO), CL (CLFO, CLO), and Vdss (VSSO).}

\item{iAUC}{interval AUC information in a dataframe with "Name", "Start", and "End" columns}

\item{excludeDelta}{Improvement of R2ADJ larger than this value could exclude the last point. Default value 1 is for the compatibility with other software. Author recommends to use \code{excludeDelta} option with about 0.3.}

\item{watermark}{Insert watermark when TRUE}

\item{internal_version}{changes temp dir pathing as a workaround for cloud hosting
where access rights prevent writing}

\item{debug_msg}{show debug messages}

\item{show_progress}{display progress messages}
}
\description{
Modified ncar::pdfNCA to support watermarks
}
