% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\name{pk_par_first_order}
\alias{pk_par_first_order}
\title{Parallel zero and first order absorption model
Adapted from https://github.com/mrgsolve/gallery/blob/master/absorption/parallel.md}
\usage{
pk_par_first_order
}
\description{
Parallel zero and first order absorption model
Adapted from https://github.com/mrgsolve/gallery/blob/master/absorption/parallel.md
}
