% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\name{pk_seq_first_order}
\alias{pk_seq_first_order}
\title{Sequential zero and first order absorption model (by Jinju Guk)}
\usage{
pk_seq_first_order
}
\description{
Sequential zero and first order absorption model (by Jinju Guk)
}
