% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{plot_data_with_nm}
\alias{plot_data_with_nm}
\title{Main plot for Simulation tab}
\usage{
plot_data_with_nm(
  input_dataset1 = NULL,
  input_dataset2 = NULL,
  nonmem_dataset = NULL,
  color_data_by = NULL,
  xvar = NULL,
  yvar = NULL,
  yvar_2 = NULL,
  log_x_axis = FALSE,
  log_x_ticks = logbreaks_x,
  log_x_labels = logbreaks_x,
  log_y_axis = FALSE,
  log_y_ticks = logbreaks_y,
  log_y_labels = logbreaks_y,
  geom_point_sim_option = FALSE,
  geom_point_data_option = FALSE,
  stat_summary_data_option = FALSE,
  stat_summary_data_by = NULL,
  nm_yvar = NULL,
  xlabel = xvar,
  ylabel = yvar,
  debug = FALSE,
  title = NULL,
  line_color_1 = "#F8766D",
  line_color_2 = "#7570B3"
)
}
\arguments{
\item{input_dataset1}{simulated dataframe for Model 1}

\item{input_dataset2}{simulated dataframe for Model 2}

\item{nonmem_dataset}{uploaded NONMEM dataset}

\item{color_data_by}{Color by a variable in the NONMEM dataset}

\item{xvar}{Name of X-variable to be plotted, string}

\item{yvar}{Name of Y-variable to be plotted, string (Model 1)}

\item{yvar_2}{Name of Y-variable to be plotted, string (Model 2)}

\item{log_x_axis, log_x_ticks, log_x_labels, log_y_axis, log_y_ticks, log_y_labels}{Log options and axis tick labels}

\item{geom_point_sim_option, geom_point_data_option}{Plotting sampling points}

\item{stat_summary_data_option}{Plots median line of dataset}

\item{stat_summary_data_by}{Inserts median line by variable}

\item{nm_yvar}{Name of Y-variable for NONMEM dataset, string}

\item{xlabel}{Nice label for X-axis, string}

\item{ylabel}{Nice label for Y-axis, string}

\item{debug}{When TRUE, displays debugging messages}

\item{title}{Title for plot}

\item{line_color_1}{Color for Model 1}

\item{line_color_2}{Color for Model 2}
}
\value{
a ggplot object
}
\description{
Main plot for Simulation tab
}
