% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_9.R
\name{plot_iiv_data_with_nm}
\alias{plot_iiv_data_with_nm}
\title{Function to plot variability data}
\usage{
plot_iiv_data_with_nm(
  input_dataset1 = NULL,
  input_dataset2 = NULL,
  nonmem_dataset = NULL,
  xvar = NULL,
  yvar = NULL,
  yvar_2 = NULL,
  log_x_axis = FALSE,
  log_x_ticks = logbreaks_x,
  log_x_labels = logbreaks_x,
  log_y_axis = FALSE,
  log_y_ticks = logbreaks_y,
  log_y_labels = logbreaks_y,
  geom_point_sim_option = FALSE,
  geom_point_data_option = FALSE,
  show_ind_profiles = FALSE,
  y_median = NULL,
  y_min = NULL,
  y_max = NULL,
  line_color_1 = "black",
  line_color_2 = "black",
  stat_summary_data_option = FALSE,
  nm_yvar = NULL,
  xlabel = xvar,
  ylabel = yvar,
  debug = FALSE,
  title = NULL,
  show_x_intercept = FALSE,
  x_intercept_value = NULL,
  show_y_intercept = FALSE,
  y_intercept_value = NULL
)
}
\arguments{
\item{input_dataset1}{simulated dataframe for Model 1}

\item{input_dataset2}{simulated dataframe for Model 2}

\item{nonmem_dataset}{uploaded NONMEM dataset}

\item{xvar}{Name of X-variable to be plotted, string}

\item{yvar}{Name of Y-variable to be plotted, string (Model 1)}

\item{yvar_2}{Name of Y-variable to be plotted, string (Model 2)}

\item{log_x_axis, log_x_ticks, log_x_labels, log_y_axis, log_y_ticks, log_y_labels}{Log options and axis tick labels}

\item{geom_point_sim_option, geom_point_data_option}{Plotting sampling points}

\item{show_ind_profiles}{If TRUE, plots individual profiles instead of prediction intervals}

\item{y_median}{Prediction interval median (when show_ind_profiles == FALSE)}

\item{y_min}{Prediction interval min (when show_ind_profiles == FALSE)}

\item{y_max}{Prediction interval max (when show_ind_profiles == FALSE)}

\item{line_color_1}{Color for Model 1 prediction intervals}

\item{line_color_2}{Color for Model 2 prediction intervals}

\item{stat_summary_data_option}{Plots median line of dataset}

\item{nm_yvar}{Name of Y-variable for NONMEM dataset, string}

\item{xlabel}{Nice label for X-axis, string}

\item{ylabel}{Nice label for Y-axis, string}

\item{debug}{When TRUE, displays debugging messages}

\item{title}{Title for plot}

\item{show_x_intercept}{Display X-intercept for threshold calculation}

\item{x_intercept_value}{Numeric for X-intercept}

\item{show_y_intercept}{Display Y-intercept for threshold calculation}

\item{y_intercept_value}{Numeric for Y-intercept}
}
\value{
a ggplot object
}
\description{
Function to plot variability data
}
