% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{plot_iiv_exp_data}
\alias{plot_iiv_exp_data}
\title{Function to plot variability exposure data}
\usage{
plot_iiv_exp_data(
  input_dataset,
  yvar = "yvar",
  ylab = yvar,
  model_1_name = "",
  model_2_name = "",
  model_1_color = "#F8766D",
  model_2_color = "#7570B3",
  show_stats = TRUE,
  xlab = "",
  title = ""
)
}
\arguments{
\item{input_dataset}{Input dataset of Model 1 and/or Model 2 that contains MODEL ID CMIN CAVG CMAX AUC}

\item{yvar}{Name of Y-variable (exposure metric) to be plotted, string}

\item{ylab}{Optional name for yvar}

\item{model_1_name}{Optional name for Model 1}

\item{model_2_name}{Optional name for Model 2}

\item{model_1_color}{Color for Model 1}

\item{model_2_color}{Color for Model 2}

\item{show_stats}{Display texts of stats for each box plot}

\item{title}{Title for plot}
}
\value{
a ggplot object
}
\description{
Function to plot variability exposure data
}
