% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{plot_three_data_with_nm}
\alias{plot_three_data_with_nm}
\title{Main function for the plot in Parameter Sensitivity Analysis}
\usage{
plot_three_data_with_nm(
  input_dataset_min,
  input_dataset_mid,
  input_dataset_max,
  param_name = "ID",
  param_min_value = -1,
  param_mid_value = -2,
  param_max_value = -3,
  x_min = NULL,
  x_max = NULL,
  nonmem_dataset = NULL,
  xvar = "TIMEADJ",
  yvar = NULL,
  log_x_axis = FALSE,
  log_x_ticks = logbreaks_x,
  log_x_labels = logbreaks_x,
  log_y_axis = FALSE,
  log_y_ticks = logbreaks_y,
  log_y_labels = logbreaks_y,
  geom_point_sim_option = FALSE,
  geom_point_data_option = FALSE,
  geom_ribbon_option = FALSE,
  geom_vline_option = FALSE,
  stat_summary_data_option = FALSE,
  nm_yvar = NULL,
  xlabel = xvar,
  ylabel = yvar,
  debug = FALSE,
  title = NULL
)
}
\arguments{
\item{input_dataset_min}{simulated dataframe for Min Parameter}

\item{input_dataset_mid}{simulated dataframe for Mid Parameter}

\item{input_dataset_max}{simulated dataframe for Max Parameter}

\item{param_name}{Name of parameter, string}

\item{param_min_value}{Min value of parameter, numeric}

\item{param_mid_value}{Mid value of parameter, numeric}

\item{param_max_value}{Max value of parameter, numeric}

\item{x_min}{Min X-axis value for time intervals to filter datasets by}

\item{x_max}{Max X-axis value for time intervals to filter datasets by}

\item{nonmem_dataset}{uploaded NONMEM dataset}

\item{xvar}{Name of X-variable to be plotted, string}

\item{yvar}{Name of Y-variable to be plotted, string}

\item{log_x_axis, log_x_ticks, log_x_labels, log_y_axis, log_y_ticks, log_y_labels}{Log options and axis tick labels}

\item{geom_point_sim_option, geom_point_data_option}{Plotting sampling points}

\item{geom_ribbon_option}{Plots AUC when TRUE}

\item{geom_vline_option}{Plots time intervals as vertical lines when TRUE}

\item{stat_summary_data_option}{Plots median line of dataset}

\item{nm_yvar}{Name of Y-variable for NONMEM dataset, string}

\item{xlabel}{Nice label for X-axis, string}

\item{ylabel}{Nice label for Y-axis, string}

\item{debug}{When TRUE, displays debugging messages}

\item{title}{Title for plot}
}
\value{
a ggplot object
}
\description{
Main function for the plot in Parameter Sensitivity Analysis
}
