% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{print_demog_plots}
\alias{print_demog_plots}
\title{Print demographics plots.}
\usage{
print_demog_plots(data)
}
\arguments{
\item{data}{input dataframe, must contain SEX (Male == 0, Female == 1),
AGE, WT}
}
\value{
a cowplot ggplot object
}
\description{
By default AGE, WT, and SEX are shown.
}
