% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{quantile_output}
\alias{quantile_output}
\title{Creates upper and lower percentiles by TIME for a given dataset}
\usage{
quantile_output(
  iiv_sim_input,
  yvar = "DV",
  lower_quartile = 0.025,
  upper_quartile = 0.975,
  dp = 5
)
}
\arguments{
\item{iiv_sim_input}{a dataframe containing TIME column}

\item{yvar}{name of column to do the quantile calculation on}

\item{lower_quartile}{lower quantile for yvar, default 0.025  (2.5\%)}

\item{upper_quartile}{lower quantile for yvar, default 0.0975 (97.5\%)}

\item{dp}{decimal places for rounding}
}
\value{
a dataframe
}
\description{
Creates upper and lower percentiles by TIME for a given dataset
}
