% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quantile_ranges_name}
\alias{quantile_ranges_name}
\title{Function that stores the names and ranges of quantiles as a
character string to be used in Figure footnotes etc (Not used)}
\usage{
quantile_ranges_name(quantiles_df, df_orig, xvar)
}
\arguments{
\item{quantiles_df}{Name of dataframe containing columns corresponding to number of quantiles}

\item{df_orig}{Name of original dataframe containing the continuous x-axis variable}

\item{xvar}{Name of x-axis variable to split by}
}
\value{
a character string containing the quantile limits
}
\description{
Function that stores the names and ranges of quantiles as a
character string to be used in Figure footnotes etc (Not used)
}
