% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_11.R
\name{quantize}
\alias{quantize}
\title{Used for dividing bins for median line plots}
\usage{
quantize(x, levels, ...)
}
\arguments{
\item{x}{vector of values (e.g. x-axis)}

\item{levels}{bins}

\item{...}{extra params to pass through to cut()}
}
\description{
Used for dividing bins for median line plots
}
