% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_12.R
\name{reconstruct_matrices}
\alias{reconstruct_matrices}
\title{Converts a matrix object into a list}
\usage{
reconstruct_matrices(
  input_model_object,
  input_matrix,
  name_of_matrix = "omega",
  coerce_to_numeric = TRUE,
  debug = show_debugging_msg
)
}
\arguments{
\item{input_model_object}{original model object that contains matrix information}

\item{input_matrix}{single collapsed matrix (assumes NA's for upper triangle and non-blocks)}

\item{name_of_matrix}{"omega" or "sigma"}

\item{coerce_to_numeric}{Default TRUE, will turn all input to numeric}

\item{debug}{When TRUE, show debugging messages}
}
\value{
a list with each element containing a block matrix
}
\description{
Each element in the list corresponds to each omega block, corresponding to provided model object
}
\note{
The original model object is needed as the configurations of matrices
are dependent on how $OMEGA and/or $SIGMA is defined in the model code.
It seems like for each repetition of $OMEGA and/or $SIGMA, a new matrix
is created.
}
