% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{safely_incProgress}
\alias{safely_incProgress}
\title{A safe version of shiny incProgress, mostly for used in testing}
\usage{
safely_incProgress(amount, detail = NULL)
}
\arguments{
\item{...}{args to be passed}
}
\value{
A shiny withProgress bar or print to console
}
\description{
A safe version of shiny incProgress, mostly for used in testing
}
\seealso{
\code{parse}
}
