% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{safely_run_single_sim}
\alias{safely_run_single_sim}
\title{purrr:safely wrappers for various functions}
\usage{
safely_run_single_sim(...)
}
\arguments{
\item{...}{args to be passed}
}
\value{
A list with two elements:
\itemize{
\item \code{result}: The result of \code{run_single_sim}, or \code{NULL} if an error occurred.
\item \code{error}: The error that occurred, or \code{NULL} if no error occurred.
}
}
\description{
purrr:safely wrappers for various functions
}
\seealso{
\code{run_single_sim}
}
