% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{safely_showNotification}
\alias{safely_showNotification}
\title{A safe version of shiny notifications, mostly for used in testing}
\usage{
safely_showNotification(...)
}
\arguments{
\item{...}{args to be passed}
}
\value{
A shiny showNotification or print to console
}
\description{
A safe version of shiny notifications, mostly for used in testing
}
\seealso{
\code{parse}
}
