% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{safely_withProgress}
\alias{safely_withProgress}
\title{A safe version of shiny withProgress, mostly for used in testing}
\usage{
safely_withProgress(message, value = 0, expr)
}
\arguments{
\item{...}{args to be passed}
}
\value{
A shiny withProgress bar or print to console
}
\description{
A safe version of shiny withProgress, mostly for used in testing
}
\seealso{
\code{parse}
}
