% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_8.R
\name{sample_age_wt}
\alias{sample_age_wt}
\title{Samples from existing databases}
\usage{
sample_age_wt(
  df_name = "None",
  nsubj = 20,
  lower.agemo = 18 * 12,
  upper.agemo = 65 * 12,
  lower.wt = 0,
  upper.wt = 100,
  prop.male = 0.5,
  seed.number = 1234
)
}
\arguments{
\item{df_name}{Name of databases, either "None", "CDC", "WHO", or "NHANES"}

\item{nsubj}{Number of subjects}

\item{lower.agemo}{Min Age (months) \emph{not applicable for "None"}}

\item{upper.agemo}{Max Age (months) \emph{not applicable for "None"}}

\item{lower.wt}{Min WT (kg)      \emph{not applicable for "None"}}

\item{upper.wt}{Max WT (kg)      \emph{not applicable for "None"}}

\item{prop.male}{Proportion of males (e.g. 0.5)  \emph{not applicable for "None"}}

\item{seed.number}{seed number}
}
\value{
a dataframe with nsubj number of rows
}
\description{
Sampling from the loaded objects \code{nhanes.filtered}, \code{who.expand}, \code{cdc.expand},
or no sampling at all ("None").
}
