% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_id_col}
\alias{search_id_col}
\title{Search for Likely ID Columns}
\usage{
search_id_col(
  orig_df,
  names_of_id_cols = c("SUBJIDN", "SUBJID", "USUBJID", "PTNO")
)
}
\arguments{
\item{orig_df}{The dataframe used for searching}

\item{names_of_id_cols}{Likely column names, in order of search priority}
}
\value{
a dataframe
}
\description{
Searches through several common column names that could be used to create the
"ID" column in the dataset, and push that to the first column of the dataset
}
