% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{smart_x_axis}
\alias{smart_x_axis}
\title{Function to improve default x-axis breaks}
\usage{
smart_x_axis(p1, max_x = NULL, xvar = "TIMEADJ", xlabel, debug = FALSE)
}
\arguments{
\item{p1}{ggplot object}

\item{max_x}{Max x-axis value, will be derived from p1 if it is set to NULL}

\item{xvar}{x variable string to be used to derive max_x}

\item{xlabel}{Will only apply when xlabel is either
"Time (hours)"
"Time (weeks)",
"Time (days)",
"Time (months)}

\item{debug}{Shows debug messages}
}
\value{
a ggplot object
}
\description{
Function to improve default x-axis breaks
}
