% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spider_plot}
\alias{spider_plot}
\title{Draws a spider plot}
\usage{
spider_plot(
  df,
  reference_value,
  param_name = ".paramname",
  ratio_name = ".ratio",
  metric_name = "Cmax",
  plot_title = "",
  display_as = "Value",
  filter_rows = 20,
  ylabname = "",
  normalize_x_axis = FALSE,
  bioeq_lines = FALSE
)
}
\arguments{
\item{df}{Input df containing parameter name, ratio (fold-change), and metric}

\item{reference_value}{Untransformed reference value}

\item{param_name}{Column name for parameters}

\item{ratio_name}{Column name for ratio}

\item{metric_name}{Column name for metric}

\item{plot_title}{Plot title}

\item{display_as}{A choice between "Ratio", "Percentage", and "Value"}

\item{filter_rows}{Filter by X number of rows}

\item{ylabname}{Custom name for y-axis}

\item{normalize_x_axis}{Treat x-axis as a factor}

\item{bioeq_lines}{When TRUE, plots the 80\% / 125\% lines relative to reference value}
}
\value{
a ggplot object
}
\description{
Draws a spider plot
}
