% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_data_frame}
\alias{split_data_frame}
\title{Function to split dataframes into roughly equal portions to facilitate
distribution to each core}
\usage{
split_data_frame(df, N = 4)
}
\arguments{
\item{df}{data.frame to split}

\item{N}{number of chunks / cores}
}
\value{
A list containing N data.frames. If N=1, then the original df
is returned unchanged
}
\description{
Function to split dataframes into roughly equal portions to facilitate
distribution to each core
}
