% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split_reference_value}
\alias{split_reference_value}
\title{Split reference values evenly between lower and upper bound for batch run table}
\usage{
split_reference_value(reference_value, lower_bound, upper_bound)
}
\arguments{
\item{reference_value}{The reference value to be used for splitting}

\item{lower_bound}{The lower bound to be used for splitting}

\item{upper_bound}{The upper bound to be used for splitting}
}
\value{
a list of splits excluding reference value
}
\description{
Split reference values evenly between lower and upper bound for batch run table
}
