% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_10.R
\name{today_numeric}
\alias{today_numeric}
\title{Get today's date as YYYYMMDD as seed input}
\usage{
today_numeric()
}
\value{
A numeric
}
\description{
Get today's date as YYYYMMDD as seed input
}
\examples{
\dontrun{
today_numeric()
}
}
