% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tornado_plot}
\alias{tornado_plot}
\title{Draws a tornado plot}
\usage{
tornado_plot(
  df,
  reference_value,
  param_name = ".paramname",
  lower_name = "Lower",
  upper_name = "Upper",
  lower_color = "#FC8D62",
  upper_color = "#66C2A5",
  metric_name = "",
  plot_title = "",
  display_as = "Percentage",
  filter_rows = 20,
  display_text = FALSE,
  xlabname = "",
  bioeq_lines = FALSE
)
}
\arguments{
\item{df}{Input df containing parameter name, lower bound name, and upper bound name}

\item{reference_value}{Untransformed reference value}

\item{param_name}{Column name for parameters}

\item{lower_name}{Column name for lower bounds}

\item{upper_name}{Column name for upper bounds}

\item{lower_color}{Fill color for lower bounds in bars}

\item{upper_color}{Fill color for upper bounds in bars}

\item{metric_name}{Exposure metric name to be used for plot label}

\item{plot_title}{Plot title}

\item{display_as}{A choice between "Ratio", "Percentage", and "Value"}

\item{filter_rows}{Filter by X number of rows}

\item{display_text}{When TRUE, displays the size as a text labl at the end of each bar}

\item{xlabname}{Custom name for X-axis (technically the Y-axis due to coord_flip)}

\item{bioeq_lines}{When TRUE, plots the 80\% / 125\% lines relative to reference value}
}
\value{
a ggplot object
}
\description{
Draws a tornado plot
}
