% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_ev_df}
\alias{transform_ev_df}
\title{Helper function to handle ev_df transformations
group}
\usage{
transform_ev_df(ev_df, model_dur, model_rate, pred_model, debug = FALSE)
}
\arguments{
\item{ev_df}{Input event dataframe}

\item{model_dur}{When TRUE, models duration}

\item{model_rate}{When TRUE, models rate}

\item{pred_model}{When TRUE, the model is a PRED model}

\item{debug}{When TRUE, outputs debugging messages}
}
\value{
a cleaned event dataframe
}
\description{
Helper function to handle ev_df transformations
group
}
