% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_14.R
\name{trim_columns}
\alias{trim_columns}
\title{Trim columns to only retain essential columns for plotting}
\usage{
trim_columns(
  data,
  x_axis,
  y_axis,
  color = "",
  type_of_plot,
  facet_name = NULL,
  insert_med_line = FALSE,
  med_line_by = "",
  ind_dose_colname = "",
  highlight_var = "",
  lloq_colname = ""
)
}
\arguments{
\item{data}{The dataframe used for trimming}

\item{x_axis}{X-axis column name used for plotting}

\item{y_axis}{Y-axis column name used for plotting}

\item{color}{(optional) Color by column}

\item{type_of_plot}{"general_plot", "ind_plot", "sim_plot"}

\item{facet_name}{(optional) column(s) to facet by}

\item{insert_med_line}{(optional) insert median line}

\item{med_line_by}{(optional) The median line to insert, insert_med_line must be TRUE}

\item{ind_dose_colname}{(optional) Individual plot dose column name}

\item{highlight_var}{(optional) highlight variable column name}

\item{lloq_colname}{(optional) LLOQ column name}
}
\value{
a dataframe
}
\description{
Useful to cut down datasets when they are large and unwieldy
}
