% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_batch_run_table}
\alias{update_batch_run_table}
\title{Update batch run parameter table}
\usage{
update_batch_run_table(
  param_df,
  lower_multiplier = 0.5,
  upper_multiplier = 1.5,
  last_change_ref_index = 0,
  length_lower_change = 0,
  length_upper_change = 0,
  show_as_character = FALSE
)
}
\arguments{
\item{param_df}{Input dataframe containing mrgsolve parameter names and values ("reference", "lower", "upper") as characters}

\item{lower_multiplier}{lower bound multiplier}

\item{upper_multiplier}{upper bound multiplier}

\item{last_change_ref_index}{Row index if last change through the UI was a reference value, otherwise 0}

\item{show_as_character}{When TRUE, coerce entire dataframe to characters}

\item{single_bound_change}{If last change was through the UI on a upper/lower bound it will be 1, otherwise will be > 1 (many bounds changed via multiplier)}
}
\value{
a df containing mrgsolve parameter names and values ("reference", "lower", "upper") as characters
}
\description{
Update batch run parameter table
}
