% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_model_object}
\alias{update_model_object}
\title{Function to update model object}
\usage{
update_model_object(input_model_object, input_new_df)
}
\arguments{
\item{input_model_object}{original model object}

\item{input_new_df}{dataframe containing new name/value combinations}
}
\value{
a mrgsolve model object
}
\description{
Function to update model object
}
