% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_resistant_popover}
\alias{update_resistant_popover}
\title{Function that adds on to the an existing bsPopover
such that the tooltip persists after updateSelectInput etc}
\usage{
update_resistant_popover(
  id,
  title,
  content,
  placement = "bottom",
  trigger = "hover",
  options = NULL
)
}
\arguments{
\item{id}{The id of the element to attach the popover to.}

\item{title}{The title of the popover.}

\item{content}{The main content of the popover.}

\item{placement}{Where the popover should appear relative to its target 
(\code{top}, \code{bottom}, \code{left}, or \code{right}). Defaults to \code{"bottom"}.}

\item{trigger}{What action should cause the popover to appear? (\code{hover},
\code{focus}, \code{click}, or \code{manual}). Defaults to \code{"hover"}.}

\item{options}{other options}
}
\value{
A popover
}
\description{
Function that adds on to the an existing bsPopover
such that the tooltip persists after updateSelectInput etc
}
