% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_7.R
\name{update_variability}
\alias{update_variability}
\title{Updates variability for a mrgmod object}
\usage{
update_variability(
  input_model_object,
  input_matrix,
  name_of_matrix = "omega",
  check_validity = TRUE,
  debug = show_debugging_msg
)
}
\arguments{
\item{input_model_object}{original model object to be updated}

\item{input_matrix}{list of matrices containing new parameter values}

\item{name_of_matrix}{"omega" or "sigma"}

\item{check_validity}{Default TRUE, where it checks that the matrix is sensible
If check fails, the model object is NOT updated}

\item{debug}{When TRUE, show debugging messages}
}
\value{
a mrgmod object
}
\description{
Updates variability for a mrgmod object
}
