% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_15.R
\name{useShinydashboardMVP}
\alias{useShinydashboardMVP}
\title{Manually define useShinydashboard before it will be removed in a future
release of shinyWidgets}
\usage{
useShinydashboardMVP()
}
\value{
Attaches shinydashboard
}
\description{
Manually define useShinydashboard before it will be removed in a future
release of shinyWidgets
}
\note{
https://github.com/dreamRs/shinyWidgets/blob/26838f9e9ccdc90a47178b45318d110f5812d6e1/R/useShinydashboard.R
}
