% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who.expand}
\alias{who.expand}
\title{World Health Organization Pediatric Data}
\format{
\subsection{\code{who.expand}}{

A data frame with 77744 rows and 6 columns SEX, AGEMO, AGEYR, WT, HT
(approx. 80k virtual subjects)
}
}
\source{
\url{https://www.who.int/childgrowth/standards/weight_for_age/en/}
}
\usage{
who.expand
}
\description{
WHO weight-for-age (https://www.who.int/childgrowth/standards/weight_for_age/en/)
from birth to 10 yo in boys and girls
Individual files are stacked together to create one overall dataframe 'who'
Insert citation notes here in the future
}
\details{
WHO length/height for age (https://www.who.int/childgrowth/standards/height_for_age/en/)
from birth to 5 yo in boys and girls
Individual files are stacked together to create one overall dataframe 'who.ht'

WHO length/height for age from birth to 5 yo in boys and girls (https://www.who.int/childgrowth/standards/height_for_age/en/)
WHO growth reference 5-19 yo in boys and girls (https://www.who.int/growthref/who2007_height_for_age/en/)
Individual files are stacked together to create one overall dataframe 'who.ht'
}
\keyword{datasets}
