% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates_v_0_2_4_exp.R
\docType{data}
\name{code_preamble}
\alias{code_preamble}
\title{Code editor preamble and postambles}
\format{
An object of class \code{character} of length 1.
}
\usage{
code_preamble
}
\description{
Code editor preamble and postambles
}
\keyword{datasets}
