% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drawDetails.watermark}
\alias{drawDetails.watermark}
\title{Auxillary function for watermark in ggplot (S3method?)}
\usage{
\method{drawDetails}{watermark}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{Other parameters passed on}
}
\description{
Auxillary function for watermark in ggplot (S3method?)
}
