% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand_addl_ii}
\alias{expand_addl_ii}
\title{Expand ADDL and II dosing rows}
\usage{
expand_addl_ii(data, x_axis, dose_col, debug = FALSE)
}
\arguments{
\item{data}{The dataframe used for expansion}

\item{x_axis}{The x_axis variable, usually TIME or TAFD etc}

\item{dose_col}{The dose variable, usually AMT or DOSE}

\item{debug}{Show debugging messages}
}
\value{
a dataframe
}
\description{
Expands ADDL and II dosing rows. If there are no ADDL and II columns, return
original dataframe unchanged
}
