% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\docType{data}
\name{model_examples_list}
\alias{model_examples_list}
\title{List of template models available for all users}
\format{
An object of class \code{character} of length 23.
}
\usage{
model_examples_list
}
\description{
List of template models available for all users
}
\keyword{datasets}
