% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates_v_0_2_15.R
\docType{data}
\name{model_examples_list}
\alias{model_examples_list}
\title{List of template models available for all users}
\format{
An object of class \code{character} of length 21.
}
\usage{
model_examples_list
}
\description{
List of template models available for all users
}
\keyword{datasets}
