% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{MVPshinyBSDep}
\alias{MVPshinyBSDep}
\title{Internal HTML Dependency for shinyBS}
\format{
A list with the following elements:
\describe{
\item{name}{Character string specifying the name of the dependency (\code{"shinyBS"}).}
\item{version}{Character string specifying the version of the dependency (\code{"0.61.1"}).}
\item{src}{A list specifying the source location of the dependency. Contains:
\itemize{
\item{\code{href}: Character string specifying the relative path (\code{"sbs"}).}
}
}
\item{meta}{\code{NULL}, indicating no metadata is provided.}
\item{script}{Character string specifying the JavaScript file (\code{"shinyBS.js"}).}
\item{stylesheet}{Character string specifying the CSS file (\code{"shinyBS.css"}).}
\item{head}{\code{NULL}, indicating no additional HTML is provided for the \verb{<head>} section.}
\item{attachment}{\code{NULL}, indicating no attachments are provided.}
\item{package}{\code{NULL}, indicating no specific package is associated.}
\item{all_files}{Logical value (\code{TRUE}) indicating that all files are included.}
}
}
\usage{
MVPshinyBSDep
}
\description{
This object represents the internal HTML dependency for the \code{shinyBS} package.
It is used to include the required JavaScript and CSS files for \code{shinyBS} functionality
without relying on unexported objects from the \code{shinyBS} package.
}
\details{
The \code{shinyBSDep} object is a list that defines the HTML dependency for \code{shinyBS}.
It includes metadata such as the name, version, source files, and other attributes.
The object is stored internally to avoid direct use of \code{shinyBS:::shinyBSDep},
which is discouraged by CRAN guidelines.
}
\note{
The \code{shinyBSDep} object is marked as internal and is not exported. It is intended
for use within the package to avoid reliance on unexported objects from \code{shinyBS}.
}
\keyword{internal}
