% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates_v_0_2_8_exp.R
\docType{data}
\name{pd_gompertz_effect}
\alias{pd_gompertz_effect}
\title{PD model of gemcitabine, with gompertz and effect compartment
Tham LS, Holford NH, Wang L, Soo RA, Lee SC, Lee HS, et al. A pharmacodynamic model for the time course of tumor shrinkage by gemcitabine + carboplatin in non-small cell lung cancer patients. Clin Cancer Res. 2008;14(13):4213-8.}
\format{
An object of class \code{character} of length 1.
}
\usage{
pd_gompertz_effect
}
\description{
PD model of gemcitabine, with gompertz and effect compartment
Tham LS, Holford NH, Wang L, Soo RA, Lee SC, Lee HS, et al. A pharmacodynamic model for the time course of tumor shrinkage by gemcitabine + carboplatin in non-small cell lung cancer patients. Clin Cancer Res. 2008;14(13):4213-8.
}
\keyword{datasets}
