% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates_v_0_2_15.R
\docType{data}
\name{pk_adaptive_dosing_int}
\alias{pk_adaptive_dosing_int}
\title{Adaptive dosing interval example from mrgsolve v1.4.1
https://mrgsolve.org/blog/posts/2024-new-1-4-0-evtools.html#update-the-dose-regimen}
\format{
An object of class \code{character} of length 1.
}
\usage{
pk_adaptive_dosing_int
}
\description{
Adaptive dosing interval example from mrgsolve v1.4.1
https://mrgsolve.org/blog/posts/2024-new-1-4-0-evtools.html#update-the-dose-regimen
}
\keyword{datasets}
