% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates_v_0_2_10.R
\docType{data}
\name{pk_seq_first_order}
\alias{pk_seq_first_order}
\title{Sequential zero and first order absorption model (by Jinju Guk)}
\format{
An object of class \code{character} of length 1.
}
\usage{
pk_seq_first_order
}
\description{
Sequential zero and first order absorption model (by Jinju Guk)
}
\keyword{datasets}
