% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\docType{data}
\name{pkpd_adaptive_dosing}
\alias{pkpd_adaptive_dosing}
\title{Adaptive dosing regimen example from mrgsolve v1.4.1
https://mrgsolve.org/blog/posts/2024-new-1-4-0-evtools.html#update-the-dose-regimen}
\format{
An object of class \code{character} of length 1.
}
\usage{
pkpd_adaptive_dosing
}
\description{
Adaptive dosing regimen example from mrgsolve v1.4.1
https://mrgsolve.org/blog/posts/2024-new-1-4-0-evtools.html#update-the-dose-regimen
}
\keyword{datasets}
