% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mvp.R
\name{run_mvp}
\alias{run_mvp}
\title{Launch Model Visualization Platform App}
\usage{
run_mvp(appDir = system.file("shiny", package = "MVPapp"), ...)
}
\arguments{
\item{appDir}{the directory of the application to run.}

\item{...}{\code{\link[shiny:runApp]{shiny::runApp()}} parameters, \code{\link[shiny:shinyApp]{shiny::shinyApp()}} parameters,
or parameters to pass to the Shiny app.}
}
\description{
This function will launch MVP app and pass along arbitrary parameters through
the \code{...} parameter to the application. This is done by modifying the
global environment. This function will attempt to clean up any objects placed
into the global environment on exit. If objects exist prior to calling this
function (i.e. \code{exists(OBJECT)} returns TRUE) then the value will be reset
to it's state prior to calling \code{run_mvp}.
}
\details{
The settings below are not technically arguments,
however when provided will toggle the app behavior:
\itemize{
\item \code{show_debugging_msg} Logical. Default \code{FALSE}. Set to \code{TRUE} to output verbose
working messages in the console, could be useful for debugging.
\item \code{authentication_code} Default \code{NULL}. Provide a string (i.e. password) to
password-lock the entire app.
\item \code{insert_watermark} Logical. Default \code{TRUE}. Set to \code{FALSE} to remove
"For Internal Use Only" text in simulated plots.
\item \code{internal_version} Logical. Default \code{TRUE}. Setting to \code{FALSE} may allow
generation of NCA reports when hosted on AWS with different access rights.
\item \code{use_bi_styling} Logical. Default \code{FALSE}. Set to TRUE to insert BI logo.
\item \code{pw_models_path} Default \code{NULL}. provide a path to source in password-gated
models (see the \code{shiny/passworded_models_example.R} for ideas, and the
vignette: \href{../doc/supply-passwords.html}{supply passworded models})
}

Note that these variables are provided internally by default and does not exist
as an object - see examples below for its correct usage.

If the user wishes to run the App outside of the function (e.g. preparing for
deployment on Posit Connect), this can be done by accessing inst/shiny/app.R,
which is located inside the folder of where the package was installed, and set
debug_mode = TRUE (and setting these options there as required).
}
\note{
Adapted from https://github.com/jbryer/ShinyDemo/blob/master/R/run_shiny_app.R
}
\examples{
\dontrun{
run_mvp(insert_watermark = FALSE) # remove watermarks
run_mvp(authentication_code = "some_password") # Password-lock the site,
# could be useful in deployment
run_mvp(pw_models_path = "path/to/your/private/models.R") # see
# "passworded_models_example.R" on how to set one up
}
}
\seealso{
\code{vignette("supply-passwords", package = "MVPapp")}
}
