% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_single_sim}
\alias{run_single_sim}
\title{Executes a mrgsim call based on user input}
\usage{
run_single_sim(
  input_model_object,
  pred_model = FALSE,
  ev_df,
  wt_based_dosing = FALSE,
  wt_name = "WT",
  model_dur = FALSE,
  model_rate = FALSE,
  sampling_times,
  seed = 1000,
  divide_by = 1,
  debug = FALSE,
  nsubj = 1,
  append_id_text = "m1-",
  ext_db = NULL,
  show_matches = FALSE,
  parallel_sim = FALSE,
  parallel_n = 200
)
}
\arguments{
\item{input_model_object}{mrgmod object}

\item{pred_model}{Default FALSE. set to TRUE to set ev_df CMT to 0}

\item{ev_df}{ev() dataframe containing dosing info}

\item{wt_based_dosing}{When TRUE, will try to use weight-based dosing}

\item{wt_name}{Name of weight parameter in the model to multiply dose amount by}

\item{model_dur}{Default FALSE, set to TRUE to model duration inside the code}

\item{model_rate}{Default FALSE, set to TRUE to model rate inside the code}

\item{sampling_times}{A vector of sampling times (note: not a tgrid object)}

\item{seed}{Seed number for RNG for reproducibility}

\item{divide_by}{Divide the TIME by this value, used for scaling x-axis}

\item{debug}{Default FALSE, set to TRUE to show more messages in console}

\item{nsubj}{Default 1, in which case mrgsolve::zero_re() will be applied}

\item{append_id_text}{A string prefix to be inserted for each ID}

\item{ext_db}{Default NULL, supply R object of external database}

\item{show_matches}{Default FALSE, shows a notification when external databases found a matching name with model parameters}

\item{parallel_sim}{Default TRUE, uses the future and mrgsim.parallel packages !Not implemented live!}

\item{parallel_n}{The number of subjects required before parallelization is used !Not implemented live!}
}
\value{
a df
}
\description{
Executes a mrgsim call based on user input
}
