% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v_0_2_13.R
\name{tblNCA_progress}
\alias{tblNCA_progress}
\title{Modified NonCompart::tblNCA to support progress bars}
\usage{
tblNCA_progress(
  concData,
  key = "Subject",
  colTime = "Time",
  colConc = "conc",
  dose = 0,
  adm = "Extravascular",
  dur = 0,
  doseUnit = "mg",
  timeUnit = "h",
  concUnit = "ug/L",
  down = "Linear",
  R2ADJ = 0,
  MW = 0,
  SS = FALSE,
  iAUC = "",
  excludeDelta = 1,
  show_progress = TRUE
)
}
\arguments{
\item{concData}{concentration data table}

\item{key}{column names of concData to be shown in the output table}

\item{colTime}{column name for time}

\item{colConc}{column name for concentration}

\item{dose}{administered dose}

\item{adm}{one of \code{"Bolus"} or \code{"Infusion"} or \code{"Extravascular"} to indicate drug administration mode}

\item{dur}{duration of infusion}

\item{doseUnit}{unit of dose}

\item{timeUnit}{unit of time}

\item{concUnit}{unit of concentration}

\item{down}{method to calculate AUC, \code{"Linear"} or \code{"Log"}}

\item{R2ADJ}{Lowest threshold of adjusted R-square value to do manual slope determination}

\item{MW}{molecular weight of drug}

\item{SS}{if steady-state, this should be TRUE. AUCLST (AUClast) is used instead of AUCIFO (AUCinf) for the calculation of Vz (VZFO, VZO), CL (CLFO, CLO), and Vdss (VSSO).}

\item{iAUC}{data.frame for interval AUC}

\item{excludeDelta}{Improvement of R2ADJ larger than this value could exclude the last point. Default value 1 is for the compatibility with other software.}

\item{show_progress}{display progress messages}
}
\description{
Modified NonCompart::tblNCA to support progress bars
}
