% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_templates.R
\name{update_model_choices}
\alias{update_model_choices}
\title{Shorthand function to update list of models and
default values by input password}
\usage{
update_model_choices(input_password, session, model_list = model_examples_list)
}
\arguments{
\item{input_password}{From the UI (input$password)}

\item{session}{Shiny session (don't change this)}

\item{model_list}{List of models (don't change this)}
}
\value{
a logical TRUE/FALSE if password is valid

\emph{Note} this function will be re-defined if passworded models is sourced externally,
i.e. when provided as an argument for run_mvp(pw_models_path = "path/to/your/private/models.R")
}
\description{
Shorthand function to update list of models and
default values by input password
}
\note{
input$model_select & input$model_select2 are updated as a side effect
}
